refactor: split git panel sections
